LayoutTests/fast/events/message-channel-gc-4-expected.txt

CONSOLE MESSAGE: line 28: ReferenceError: Can't find variable: MessageChannel
FAIL: Timed out waiting for notifyDone to be called
Test that MessagePort messages are delivered even if both ports are in transit (cloned).

Should say PASS twice.
